% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affordability_matrix_from_chas.R
\name{get_affordability_matrix_from_chas}
\alias{get_affordability_matrix_from_chas}
\title{Get Matrix of Household Income and Housing Affordability (CHAS Data 2013-2017)}
\usage{
get_affordability_matrix_from_chas(
  geography,
  state,
  bin_adjustment_type = "unify"
)
}
\arguments{
\item{geography}{Options are "county" or "place", determines the geographic unit to return CHAS data for.}

\item{state}{U.S. state to return data for, requires the state two letter abbreviation to be passed}

\item{bin_adjustment_type}{Options are either "unify" or "low end ownership", if "unify" is selected the bins are unified across renters and owners which makes the results less granular but no assumptions are made. "low end ownership" preserves the 0-30% affordability bin but assumes all ownership units in 0-50% are actuall in 30-50%.}
}
\description{
This function formats table 18 of the 2013-2017 CHAS data to return the "matrix" of units by household income and unit affordability.
}
\examples{
get_affordability_matrix_from_chas(geography = "place", state = "ID", bin_adjustment_type = "unify")
}
