% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_income_and_rent_for_renters_from_chas.R
\name{get_income_and_rent_for_renters_from_chas}
\alias{get_income_and_rent_for_renters_from_chas}
\title{Get Count of Housing Units by Income and Rent for Renters (CHAS Data 2013-2017)}
\usage{
get_income_and_rent_for_renters_from_chas(geography, state)
}
\arguments{
\item{geography}{Options are "county" or "place", determines the geographic unit to return CHAS data for.}

\item{state}{U.S. state to return data for, requires the state two letter abbreviation to be passed}
}
\description{
This function formats table 18C of the 2013-2017 CHAS data to return the count of units for renters by income and affordability.
}
\examples{
get_income_and_rent_for_renters_from_chas(geography = "place", state = "ID")
}
