% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_income_and_value_for_owners_without_mortgage_from_chas.R
\name{get_income_and_value_for_owners_without_mortgage_from_chas}
\alias{get_income_and_value_for_owners_without_mortgage_from_chas}
\title{Get Count of Housing Units by Income and Value of Home for Owners without a Mortgage (CHAS Data 2013-2017)}
\usage{
get_income_and_value_for_owners_without_mortgage_from_chas(geography, state)
}
\arguments{
\item{geography}{Options are "county" or "place", determines the geographic unit to return CHAS data for.}

\item{state}{U.S. state to return data for, requires the state two letter abbreviation to be passed}
}
\description{
This function formats table 18B of the 2013-2017 CHAS data to return the count of units for owners without a mortgage by income and affordability.
}
\examples{
get_income_and_value_for_owners_without_mortgage_from_chas(geography = "county", state = "ID")
}
