% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_end_ownership_rent_and_value.R
\name{low_end_ownership_rent_and_value}
\alias{low_end_ownership_rent_and_value}
\title{Internal function to unify rent and ownership CHAS bins}
\usage{
low_end_ownership_rent_and_value(renters_data, owners_data)
}
\description{
Internal function to unify rent and ownership CHAS bins
}
